ViPER engine: full RE verification against decompiled v2505 + FET/VHE
Verified every unit (engine chain, effects, SoftwareLimiter, FETCompressor, shared biquad/filter
primitives, Convolver, Equalizer) against the Ghidra decompilation. Extracted ~150 constants directly
from the binary (.text literal pools /.rodata via a Ghidra data dump) — no guessed coefficients; each
cited by address.

Reimplemented two fictional effects to match the original: FIREqualizer (was an FFT stub with the
wrong API) and PlaybackGain/AGC (was an ad-hoc heuristic; recovered from disassembly where the
decompiler lost the NEON float math). Implemented the two missing effects FET Compressor + VHE
(Headphone Surround+), wired through the engine/JNI/Kotlin.

Fixed real fidelity bugs: ViPERClarity/NoiseSharpening feedback, ViPERBass x4 (incl. SUBWOOFER
NaN), ColorfulMusic/DepthSurround, SpectrumExtend, Reverberation, the effect-chain order, and
cross-unit (Reverberation::SetSamplingRate). Accurate variable names + ~80 TODOs resolved. Native
build green.

NOTE: changes DSP/audio behavior — audition on-device. Reverse-engineered; fully revertible.

+++ app/src/main/cpp/viper/utils/DynamicBass.h
@@ -25,7 +25,7 @@ public:
     uint32_t lowFreqX, highFreqX;
     uint32_t lowFreqY, highFreqY;
     uint32_t samplingRate;
-    float qPeak;
+    int32_t qPeak;   // truncated, clamped to <= 1600 (binary stores it as an int)
     float bassGain;
     float sideGainX, sideGainY;
